Fixed various commenting and logical errors, declared combine() and operator+ functions, and created Main file
I shifted "public:" down below "private" to declare the operator+ function. If you guys can think of something better, feel free to change it! It also goes with any of the other functions I declared/ initialized. FYI: the set_polynomial() function worked for me in both test cases in the Project Instructions. However, if you can find a test case where it doesn't work, please let me know! I'll be happy to try and debug the function. As for the set_polynomial() function not directly taking in a string: I found that when I tried declaring the function this way [void set_polynomial(std::string poly)], it returned continuous errors as it kept making the "-" string a "?". I couldn't find a way to fix this, so I made the user define the polynomial string in the console once the function is called.